+++ gatsby-config.js
@@ -40,9 +40,9 @@ module.exports = {
 				accessToken: process.env.PRISMIC_ACCESS_TOKEN,
 				repositoryName: process.env.PRISMIC_REPOSITORY_NAME,
 				schemas: {
-					homepage: require("./src/.schemas/homepage.schema.json"),
-					page: require("./src/.schemas/page.schema.json"),
-					post: require("./src/.schemas/post.schema.json"),
+					homepage: require("./src/schemas/homepage.schema.json"),
+					page: require("./src/schemas/page.schema.json"),
+					post: require("./src/schemas/post.schema.json"),
 				},
 				shouldDownloadImage: () => true,
 			},
