+++ jest.config.js
@@ -7,11 +7,11 @@ module.exports = {
 		".+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$":
 			"<rootDir>/__mocks__/file-mock.js",
 	},
-	setupFiles: ["<rootDir>/src/.testing/loadershim.js"],
-	setupFilesAfterEnv: ["<rootDir>/src/.testing/setup-test-env.js"],
+	setupFiles: ["<rootDir>/src/testing/loadershim.js"],
+	setupFilesAfterEnv: ["<rootDir>/src/testing/setup-test-env.js"],
 	testPathIgnorePatterns: ["node_modules", "\\.cache", "<rootDir>.*/public"],
 	transform: {
-		"^.+\\.[jt]sx?$": "<rootDir>/src/.testing/jest-preprocess.js",
+		"^.+\\.[jt]sx?$": "<rootDir>/src/testing/jest-preprocess.js",
 	},
 	transformIgnorePatterns: ["node_modules/(?!(gatsby)/)"],
 };
