+++ src/components/Menu.tsx
@@ -1,12 +1,12 @@
 import { StaticQuery, graphql } from "gatsby";
-import { StoreActions, StoreState } from "../.types/store.types";
+import { StoreActions, StoreState } from "../types/store.types";
 import { getLanguageCode, getSlug } from "../utils/localization.util";
 import { Button } from "./Button";
 import { Context } from "./Context";
 import { LangSwitcher } from "./LangSwitcher";
 import { Link } from "./Link";
 import { MenuIcon } from "./Icon";
-import { PrismicContext } from "../.types/prismic.types";
+import { PrismicContext } from "../types/prismic.types";
 import React from "react";
 import { connect } from "react-redux";
 import { getTranslation } from "../utils/translation.util";
