+++ src/types/component.types.ts
